Removes dependencies of {EmailTemplateProcessor, FriendRecommendationsEmailTip} on UserCommander
