+++ kifi-backend/modules/shoebox/test/com/keepit/common/store/FakeS3ImageStore.scala
@@ -10,8 +10,8 @@ import scala.concurrent._
 import scala.util.{ Success, Try }
 
 case class FakeS3ImageStore(val config: S3ImageConfig) extends S3ImageStore {
-  def getPictureUrl(w: Int, user: User) =
-    Promise[String]().success(s"//cloudfront/${user.id.get}_${w}x${w}").future
+  def getPictureUrl(w: Int, userId: Id[User]) =
+    Promise[String]().success(s"//cloudfront/${userId}_${w}x${w}").future
   def getPictureUrl(width: Option[Int], user: User, picVersion: String): Future[String] =
     Promise[String]().success(s"//cloudfront/users/${user.id.get}/pics/${width.getOrElse(100)}/$picVersion.jpg").future
